Extract notifications into new package
