+++ apps/app/.env.example
@@ -30,5 +30,3 @@ NEXT_PUBLIC_POSTHOG_HOST=""
 NEXT_PUBLIC_APP_URL="http://localhost:3000"
 NEXT_PUBLIC_WEB_URL="http://localhost:3001"
 NEXT_PUBLIC_DOCS_URL="http://localhost:3004"
-NEXT_PUBLIC_KNOCK_FEED_CHANNEL_ID=""
-NEXT_PUBLIC_KNOCK_API_KEY=""
