--- packages/design-system/components/knock-notification-feed.tsx
@@ -1,30 +1,0 @@
-'use client';
-
-import { useState, useRef } from "react";
-import {
-  NotificationIconButton,
-  NotificationFeedPopover,
-} from "@knocklabs/react";
-
-// Required CSS import, unless you're overriding the styling
-import "@knocklabs/react/dist/index.css";
-
-export const KnockNotificationFeed = () => {
-  const [isVisible, setIsVisible] = useState(false);
-  const notifButtonRef = useRef(null);
-
-  return (
-
-        <>
-          <NotificationIconButton
-            ref={notifButtonRef}
-            onClick={(e) => setIsVisible(!isVisible)}
-          />
-          <NotificationFeedPopover
-            buttonRef={notifButtonRef}
-            isVisible={isVisible}
-            onClose={() => setIsVisible(false)}
-          />
-        </>
-  );
-};
