--- packages/design-system/lib/knock.ts
@@ -1,9 +1,0 @@
-import { Knock } from '@knocklabs/node';
-
-const knockApiKey = process.env.KNOCK_SECRET_API_KEY;
-
-if (!knockApiKey) {
-  throw new Error('KNOCK_SECRET_API_KEY is not set');
-}
-
-export const knock = new Knock(knockApiKey);
