+++ packages/design-system/package.json
@@ -8,8 +8,6 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^3.9.1",
-    "@knocklabs/node": "^0.6.13",
-    "@knocklabs/react": "^0.2.29",
     "@radix-ui/react-accordion": "^1.2.2",
     "@radix-ui/react-alert-dialog": "^1.1.4",
     "@radix-ui/react-aspect-ratio": "^1.1.1",
