+++ packages/notifications/index.ts
@@ -1,0 +1,6 @@
+import { Knock } from '@knocklabs/node';
+import { keys } from './keys';
+
+const key = keys().KNOCK_SECRET_API_KEY;
+
+export const notifications = new Knock(key);
